Update : Directory by aglrithm type
