+++ backtracking/BAEKJOON_15650.java
@@ -1,8 +1,6 @@
-//package baekjoon;
-
 import java.util.Scanner;
 import java.util.Stack;
-
+//N과 M (2)
 public class BAEKJOON_15650 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
