--- baekjoon/BAEKJOON_24262.java
@@ -1,12 +1,0 @@
-//package baekjoon;
-
-import java.util.Scanner;
-
-public class BAEKJOON_24262 {
-	public static void main(String[] args) {
-		Scanner sc = new Scanner(System.in);
-		int n = sc.nextInt();
-		
-		System.out.printf("%d\n%d",1, 0);
-	}
-}
