--- baekjoon/BAEKJOON_25314.java
@@ -1,12 +1,0 @@
-//package baekjoon;
-
-import java.util.Scanner;
-
-public class BAEKJOON_25314 {
-	public static void main(String[] args) {
-		Scanner sc = new Scanner(System.in);
-		int n = sc.nextInt();
-		String l = "long ";
-		System.out.println(l.repeat(n/4) +"int");
-	}
-}
