+++ binary_search/BAEKJOON_10815.java
@@ -1,5 +1,3 @@
-//package baekjoon;
-
 import java.util.Arrays;
 import java.util.Scanner;
