+++ greedy/BAEKJOON_1083.java
@@ -1,4 +1,3 @@
-//package baekjoon;
 /*
  * 
  * 1 2 3 4 5 6 7 8 9 10
