+++ greedy/BAEKJOON_11047.java
