+++ greedy/BAEKJOON_11399.java
@@ -1,4 +1,3 @@
-//package baekjoon;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Scanner;
