+++ greedy/BAEKJOON_13305.java
