+++ implementation/BAEKJOON_10951.java
@@ -1,5 +1,3 @@
-//package baekjoon;
-
 import java.util.Scanner;
 
 public class BAEKJOON_10951 {
