+++ implementation/BAEKJOON_11654.java
@@ -1,4 +1,3 @@
-//package baekjoon;
 //string;
 import java.util.*;
