+++ implementation/BAEKJOON_15552.java
@@ -1,5 +1,3 @@
-//package baekjoon;
-
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.IOException;
