+++ implementation/BAEKJOON_15596.java
@@ -1,5 +1,3 @@
-//package baekjoon;
-
 import java.util.Arrays;
 
 public class BAEKJOON_15596 {
