+++ implementation/BAEKJOON_25305.java
@@ -1,5 +1,3 @@
-//package baekjoon;
-
 import java.io.*;
 import java.util.StringTokenizer;
