+++ string/BAEKJOON_1316.java
@@ -1,4 +1,3 @@
-//package baekjoon;
 //string;
 
 import java.io.BufferedReader;
