+++ string/BAEKJOON_1427.java
@@ -1,5 +1,3 @@
-// package baekjoon;
-
 import java.io.*;
 import java.util.Arrays;
