+++ string/BAEKJOON_14425.java
@@ -1,5 +1,3 @@
-//package baekjoon;
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
